BCs as subclasses. Part1: define
